requirements.txt
fastapi
mangum
uvicorn
dotenv
email-validator
python-jose[cryptography]
boto3
autopep8